+++ src/main/java/com/arangodb/internal/InternalArangoDatabase.java
@@ -26,7 +26,7 @@ import com.arangodb.entity.arangosearch.analyzer.SearchAnalyzer;
 import com.arangodb.internal.ArangoExecutor.ResponseDeserializer;
 import com.arangodb.internal.util.ArangoSerializationFactory;
 import com.arangodb.internal.util.ArangoSerializationFactory.Serializer;
-import com.arangodb.internal.util.EncodeUtils;
+import com.arangodb.util.UnicodeUtils;
 import com.arangodb.internal.util.RequestUtils;
 import com.arangodb.model.*;
 import com.arangodb.model.arangosearch.AnalyzerDeleteOptions;
@@ -80,7 +80,7 @@ public abstract class InternalArangoDatabase<A extends InternalArangoDB<EXECUTOR
         super(arango.executor, arango.util, arango.context);
         this.arango = arango;
         this.name = name;
-        EncodeUtils.checkNormalized(name);
+        UnicodeUtils.checkNormalized(name);
     }
 
     public A arango() {
